fix syntax error for delete_security_grant

+++ src/main/kotlin/org/seedcompany/savor/core/data/sql/security/delete/delete_security_grant.sql
@@ -20,7 +20,7 @@ begin
 				raise info 'new_access_level: %', new_access_level;
                 execute format('update '|| security_table_name ||' set '|| security_column_name || ' = NULL  where __sys_person_id = '|| rec1.sys_person_id );
             else 
-                execute format('update '|| security_table_name ||' set '|| security_column_name ' = '|| new_access_level || ' where __sys_person_id = ' || rec1.sys_person_id);
+                execute format('update '|| security_table_name ||' set '|| security_column_name || ' = '|| new_access_level || ' where __sys_person_id = ' || rec1.sys_person_id);
             end if;
 		end loop;
 		raise info 'done';
@@ -35,7 +35,7 @@ ON sys_role_grants
 FOR EACH ROW
 EXECUTE PROCEDURE delete_security_on_grant_fn();
 
-delete from sys_role_grants;
+delete from sys_role_grants where  column_name = 'name' and access_level = 'Write';
 select * from sys_add_role_grant('Admin', 'Seed Company', 'sys_locations', 'name', 'Write');
 select * from sys_role_grants;
